requirements.txt
numpy==1.21.6
opencv_python==4.8.1.78
Pillow==9.5.0
Pillow==11.0.0
scikit_image==0.19.3
scikit_learn==1.0.2
timm==0.9.10
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.1
